requirements.txt
langchain==0.1.3
langchain-community==0.0.16
huggingface-hub==0.20.2
transformers==4.36.2
sentence-transformers==2.2.2
numpy==1.26.3
qdrant-client==1.7.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
tiktoken==0.5.2
pypdf==3.17.1
streamlit==1.29.0
torch==2.1.2 